Plotly/requirements.txt
numpy==2.4.2
pandas==2.3.3
dash==4.0.0
plotly==6.5.2
google-cloud-bigquery==3.40.1
db-dtypes==1.5.0
gunicorn==25.1.0